rework project creation modal
