+++ packages/legend-art/style/base/_utils.scss
@@ -14,14 +14,6 @@
  * limitations under the License.
  */
 
-.u-full-height {
-  height: 100%;
-}
-
-.u-full-width {
-  width: 100%;
-}
-
 .u-pull-right {
   float: right;
 }
