+++ packages/legend-query/style/_query-builder.scss
@@ -613,6 +613,10 @@
   height: 100%;
   width: 100%;
 
+  &__enum-selector {
+    width: 100%;
+  }
+
   ::-webkit-calendar-picker-indicator {
     filter: invert(0.8);
   }
