+++ packages/legend-studio/style/components/_create-project-modal.scss
@@ -1,0 +1,177 @@
+/**
+ * Copyright (c) 2020-present, Goldman Sachs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+@use 'mixins' as *;
+
+.setup__create-workspace-modal {
+  width: 75rem;
+  border: 0.1rem solid var(--color-blue-200);
+  padding: 2rem;
+  overflow: auto;
+
+  &__doc__create-workspace {
+    margin-left: 0.5rem;
+    color: var(--color-dark-grey-400);
+  }
+
+  &__form {
+    display: flex;
+    padding: 0.5rem 0 1rem;
+    flex-direction: column;
+
+    &__workspace {
+      flex-direction: column;
+    }
+
+    &__project-name__input {
+      flex: 1 0 auto;
+      background: var(--color-dark-grey-280);
+      color: var(--color-light-grey-200);
+      border-radius: 0 0.2rem 0.2rem 0;
+      height: 3rem;
+      padding: 1rem;
+    }
+
+    &__workspace-name__input {
+      flex: 1 0 auto;
+      background: var(--color-dark-grey-280);
+      color: var(--color-light-grey-200);
+      border-radius: 0 0.2rem 0.2rem 0;
+      height: 3rem;
+      padding: 1rem;
+      margin-top: 1rem;
+    }
+  }
+}
+
+.setup__create-project-modal {
+  width: 75rem;
+  border: 0.1rem solid var(--color-blue-200);
+  padding: 0;
+  overflow: auto;
+
+  &__header {
+    display: flex;
+    margin: 1.5rem 2rem;
+
+    &__label {
+      font-size: 2rem;
+      font-weight: bold;
+    }
+
+    &__action {
+      color: var(--color-blue-200);
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      width: 2.8rem;
+
+      .svg {
+        height: 2rem;
+      }
+    }
+
+    &__tabs {
+      height: 2.8rem;
+      z-index: 1;
+      display: flex;
+      overflow-x: overlay;
+      overflow-y: hidden;
+      background: var(--color-dark-grey-80);
+    }
+
+    &__tab {
+      @include flexCenter;
+
+      display: inline-flex;
+      height: 100%;
+      color: var(--color-light-grey-400);
+      padding: 0 2rem;
+      border-right: 0.1rem solid var(--color-dark-grey-100);
+      white-space: nowrap;
+      cursor: pointer;
+    }
+
+    &__tab__doc {
+      margin-left: 0.5rem;
+      color: var(--color-dark-grey-400);
+    }
+
+    &__tab--active {
+      position: relative;
+    }
+
+    &__tab--active::after {
+      content: '';
+      height: 0.2rem;
+      width: 100%;
+      position: absolute;
+      bottom: 0;
+      background: var(--color-yellow-200);
+    }
+  }
+
+  &__container {
+    overflow-y: hidden !important;
+  }
+
+  &__content {
+    padding: 2rem;
+  }
+
+  &__inner-container {
+    overflow-y: initial !important;
+  }
+
+  .selector-input__menu {
+    z-index: 1;
+  }
+
+  &__submit-btn {
+    height: 3.6rem;
+  }
+
+  &__success {
+    width: 100%;
+    background: var(--color-green-80);
+    color: var(--color-dark-grey-50);
+
+    &__label {
+      line-height: 1.8rem;
+      padding: 0.5rem;
+      font-size: 1.6rem;
+
+      &__link {
+        padding-left: 0.3rem;
+        padding-right: 0.3rem;
+        color: var(--color-dark-grey-50);
+      }
+    }
+  }
+
+  &__form {
+    display: flex;
+    padding: 0.5rem 0 1rem;
+    flex-direction: column;
+
+    &__unsupported {
+      @include flexHCenter;
+
+      font-weight: 500;
+      color: var(--color-dark-grey-400);
+    }
+  }
+}
