--- packages/legend-studio/style/components/_create-project.scss
@@ -1,109 +1,0 @@
-/**
- * Copyright (c) 2020-present, Goldman Sachs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-.setup-create {
-  &__modal {
-    width: 75rem;
-    border: 0.1rem solid var(--color-blue-200);
-    padding: 2rem;
-    overflow: auto;
-
-    &__heading {
-      display: flex;
-      margin-bottom: 1rem;
-
-      &__label {
-        font-size: 2rem;
-        font-weight: bold;
-      }
-
-      &__action {
-        color: var(--color-blue-200);
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        width: 2.8rem;
-
-        .svg {
-          height: 2rem;
-        }
-      }
-    }
-
-    &__container {
-      overflow-y: hidden !important;
-    }
-
-    &__inner-container {
-      overflow-y: initial !important;
-    }
-
-    .selector-input__menu {
-      z-index: 1;
-    }
-
-    &__submit-btn {
-      height: 3.6rem;
-    }
-
-    &__success {
-      width: 100%;
-      background: var(--color-green-80);
-      color: var(--color-dark-grey-50);
-
-      &__label {
-        line-height: 1.8rem;
-        padding: 0.5rem;
-        font-size: 1.6rem;
-
-        &__link {
-          padding-left: 0.3rem;
-          padding-right: 0.3rem;
-          color: var(--color-dark-grey-50);
-        }
-      }
-    }
-  }
-
-  &__form {
-    display: flex;
-    padding: 0.5rem 0 1rem;
-    flex-direction: column;
-
-    &__workspace {
-      flex-direction: column;
-    }
-
-    &__project-name__input {
-      flex: 1 0 auto;
-      background: var(--color-dark-grey-280);
-      color: var(--color-light-grey-200);
-      border-radius: 0 0.2rem 0.2rem 0;
-      height: 3rem;
-      padding: 1rem;
-    }
-
-    &__workspace-name__input {
-      flex: 1 0 auto;
-      background: var(--color-dark-grey-280);
-      color: var(--color-light-grey-200);
-      border-radius: 0 0.2rem 0.2rem 0;
-      height: 3rem;
-      padding: 1rem;
-      margin-top: 1rem;
-    }
-  }
-}
