+++ packages/legend-studio/style/components/_setup.scss
@@ -16,6 +16,8 @@
 
 @use 'mixins' as *;
 
+@forward './create-project-modal';
+
 .setup {
   height: 100%;
   position: relative;
@@ -104,12 +106,6 @@
     font-size: 2rem;
     color: var(--color-dark-grey-400);
     margin-left: 1rem;
-  }
-
-  &__doc__create-project,
-  &__doc__create-workspace {
-    margin-left: 0.5rem;
-    color: var(--color-dark-grey-400);
   }
 
   &__workspace__label {
