+++ packages/legend-studio/style/index.scss
@@ -21,4 +21,3 @@
 @forward 'components/editor';
 @forward 'components/viewer';
 @forward 'components/review';
-@forward 'components/create-project';
